Move `compare_arrays`.
